Create separate files for dev demo components
