+++ index.html
@@ -9,6 +9,6 @@
 
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/dev/AppDev.tsx"></script>
+    <script type="module" src="/src/dev/index.tsx"></script>
   </body>
 </html>
